We can't use `depth` directly, since we need a value in `[0..1]` range. Also it'd be nice to have a smooth "gradient" like fog. We can apply glsl [smoothstep](https://thebookofshaders.com/glossary/?search=smoothstep) function to calcuate the final amount of fog. This function interpolates a value in range of `lowerBound` and `upperBound`. Max depth of our camera is `142`
```javascript
mat4.perspective(
    projectionMatrix,
    (Math.PI / 360) * 90,
    canvas.width / canvas.height,
    0.01,
    142 // <- zFar
);
```

So the max value of `depth` should be < 142 in order to see any fog at all (object farther than 142 won't be visible at all). Let's use `60..100` range.

One more thing to take into account is that we don't want to see the object _completely_ white, so let's multiply the final amount by `0.9`

We'll need the final value of `fogAmount` in fragment shader, so this should be a `varying`